WISP.data: Managing WISPstation Hyperspectral Data

R package designed to automate the acquisition, quality control, analysis, and visualization of spectral data collected by the WISPstation fixed spectroradiometer.

See also

Author

Maintainer: Alessandro Oggioni alessandro.oggioni@cnr.it (ORCID) [funder]

Authors:

Nicola Ghirardi nicola.ghirardi@cnr.it (ORCID)