SUNGLINT Removal (SR) for WISPstation reflectance data

This function applies the algorithm of Jiang et al., (2020) for removing sunglint from spectral signatures

Usage

wisp_sr_reflectance_data(
  qc_data,
  calc_scatt = TRUE,
  calc_SPM = TRUE,
  calc_TUR = TRUE,
  calc_TSS = TRUE,
  calc_gons = TRUE,
  calc_gons740 = TRUE,
  calc_NDCI = TRUE,
  calc_mishra = TRUE,
  save_csv = FALSE,
  out_dir = "outputs"
)

Arguments

qc_data

A tibble from wisp_qc_reflectance_data() function.

calc_scatt

A logical. If TRUE, the function calculates the peak due to phytoplankton scattering (690-710 nm) and the ratio of the latter to the second chlorophyll absorption peak (670-680 nm). Default is TRUE.

calc_SPM

A logical. If TRUE, the function calculates the SPM concentrations in according to Novoa et al., (2017). Default is TRUE.

calc_TUR

A logical. If TRUE, the function calculates the turbidity (FNU) in according to Novoa et al., (2017). Default is TRUE.

calc_TSS

A logical. If TRUE, the function calculates the TSS concentrations in according to Jiang et al., (2021). Default is TRUE.

calc_gons

A logical. If TRUE, the function calculates chlorophyll using Gons et al. (2002) algorithm (NIR ~782 nm). Default is TRUE.

calc_gons740

A logical. If TRUE, the function calculates chlorophyll using Gons et al. (2002) algorithm (NIR ~740 nm). Default is TRUE.

calc_NDCI

A logical. If TRUE, the function calculates The Normalised Difference Chlorophyll Index algorithm by Mishra and Mishra (2012). Default is TRUE.

calc_mishra

A logical. If TRUE, the function calculates chlorophyll using Mishra and Mishra (2012) algorithm. Default is TRUE.

save_csv

A logical. If TRUE, the function saves the reflectance data.

out_dir

A character. The directory where the CSV file will be saved. Default is "outputs" within the working directory.

Value

A tibble with the spectral signatures after the SR operation and, if parameter save_out_sr is TRUE, the function saves the reflectance data in a CSV file.

Author

Alessandro Oggioni, phD alessandro.oggioni@cnr.it

Nicola Ghirardi, phD nicola.ghirardi@cnr.it

Examples

# example code
if (FALSE) { # \dontrun{
## Not run:
reflect_data_sr <- wisp_sr_reflectance_data(
  qc_data = reflect_data_qc,
  calc_scatt = TRUE,
  calc_SPM = TRUE,
  calc_TUR = TRUE,
  calc_TSS = TRUE,
  calc_gons  = TRUE,
  calc_gons740 = TRUE,
  calc_NDCI = TRUE,
  calc_mishra = FALSE,
  save_csv = FALSE,
  out_dir = "outputs"
)
} # }
## End (Not run)